refactor: split helper function from main.js
