+++ 12_Calendar_Datepicker/index.html
@@ -40,6 +40,6 @@
         </div>
       </div>
     </div>
-    <script src="main.js"></script>
+    <script type="module" src="main.js"></script>
   </body>
 </html>
