+++ 12_Calendar_Datepicker/style.css
@@ -15,6 +15,7 @@ body {
   font-family: 'Open Sans';
   font-weight: 300;
   text-align: center;
+  min-height: 100vh;
 }
 
 .title {
